+++ isis2304-tutorial-arquitectura/src/dao/DAOServicio.java
@@ -100,7 +100,7 @@ public class DAOServicio
 	 * @throws SQLException SQLException Genera excepcion si hay error en la conexion o en la consulta SQL
 	 * @throws Exception Si se genera un error dentro del metodo.
 	 */
-	public void deleteServicio(Integer id) throws SQLException, Exception 
+	public void deleteServicio(Long id) throws SQLException, Exception 
 	{
 
 		String sql = String.format("DELETE FROM %1$s.SERVICIOS WHERE ID = %2$d", USUARIO, id);
